FEUniqueID.h will no longer include windows.h; FETime now supports non-Windows systems; FEThreadPool will use std::sleep instead of Sleep; FEThreadPool will skip worker thread creation on Emscripten builds as the current implementation is not compatible; FELog will use portable std::thread; FEConsoleWindow will not be compiled under Emscripten; Added platform files for Emscripten; Fixed issues with WebGPU under Emscripten target; CMake will now enforce the correct platform and WebGPU settings when targeting Emscripten; Network functionality is excluded from Emscripten builds; Dawn libraries are no longer used directly when targeting Emscripten; Fixed deprecated '-sUSE_WEBGPU=1' flag; Application should route the main loop through FEPlatformInterface::RunMainLoop to ensure compatibility with web deployment.

+++ FELog.h
@@ -21,7 +21,7 @@ namespace FocalEngine
 		int Count;
 		std::string Topic;
 		uint64_t TimeStamp;
-		DWORD ThreadID;
+		size_t ThreadID;
 
 		LogItem();
 		~LogItem();
